refactor: move FormalStreamExtender
